Move CheckState to data/
